Error free in 2023
You know what. It's 2025. 2 Years since 2023, a year a could have said that. 365 days ripe for that banger. Covid was ending. 3 years before that covid started. A year prior... 6 years. Hmm. Anyways, this is finally in kinda a state of 100 working. No errors everything "works" as intended. However a lot of work to go into the logic for walking, breaking up the code and whatnot.