slime attacks and health goes up
next would be the transition to the red die.
But also there's a consideration here... Do we want to keep health as a float? Or should it be treated as an int... I was thinking of float to be able to do any math easier and future proof it, but there may be so many uglinesses if I stay with float... hmmm