MILESTONE: mob die functional
however, when we roll a WOLF there is not wolf mob yet. that requires a few things: reorganizing knight, mob, slime classes and any related logic. Then create wolf mob, then create logic for dealing a card with a wolf instead of knight.
Once we have this, we can deal a card with a bigger knight to represent the double knight. THEN we can implement the double knight summon (i.e. 2 knights on invoke == 1 big knight with 3x damage)
(also to impliment: burn card when adding too many to invoke)